Major refactoring. ZIP files deletion. Self homology in ascending order. Add bp unit to measure the self homology.
