+++ .gitignore
@@ -3,3 +3,4 @@ target
 .idea
 antiSMASH-OOPS.iml
 lifeTreeOutput
+selfhomology
