--- src/main/java/uk/ac/mib/antismashoops/core/service/params/KnownClusters.java
@@ -1,6 +1,0 @@
-package uk.ac.mib.antismashoops.core.service.params;
-
-public class KnownClusters
-{
-
-}
