+++ src/main/java/uk/ac/mib/antismashoops/web/utils/WorkspaceManager.java
@@ -1,7 +1,6 @@
 package uk.ac.mib.antismashoops.web.utils;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Date;
@@ -74,22 +73,6 @@ public class WorkspaceManager
 
     public void deleteWorkspace(String wsName) throws IOException
     {
-        this.delete(new File(this.getRootDirectory(), wsName));
-    }
-
-
-    public void delete(File f) throws IOException
-    {
-        if (f.isDirectory())
-        {
-            for (File c : f.listFiles())
-            {
-                delete(c);
-            }
-        }
-        if (!f.delete())
-        {
-            throw new FileNotFoundException("Failed to delete file: " + f);
-        }
+        FileUtils.delete(new File(this.getRootDirectory(), wsName));
     }
 }
